--- a/.travis.yml
+++ b/.travis.yml
@@ -1,16 +1,13 @@
 language: go
 sudo: false
 go:
-  - 1.2
-  - 1.3
-  - 1.4
-  - 1.5
-  - 1.6
+  - 1.7
+  - 1.8
+  - 1.9
   - tip
 
 before_install:
   - go get github.com/mattn/goveralls
-  - go get golang.org/x/tools/cmd/cover
 
 script:
   - goveralls -service=travis-ci