--- a/stack.go
+++ b/stack.go
@@ -1,3 +1,5 @@
+// +build go1.7
+
 // Package stack implements utilities to capture, manipulate, and format call
 // stacks. It provides a simpler API than package runtime.
 //
@@ -21,23 +23,31 @@ import (
 
 // Call records a single function invocation from a goroutine stack.
 type Call struct {
-	pcs [2]uintptr
+	frame runtime.Frame
 }
 
 // Caller returns a Call from the stack of the current goroutine. The argument
 // skip is the number of stack frames to ascend, with 0 identifying the
 // calling function.
 func Caller(skip int) Call {
-	var pcs [2]uintptr
+	// As of Go 1.9 we need room for up to three PC entries.
+	//
+	// 0. An entry for the stack frame prior to the target to check for
+	//    special handling needed if that prior entry is runtime.sigpanic.
+	// 1. A possible second entry to hold metadata about skipped inlined
+	//    functions. If inline functions were not skipped the target frame
+	//    PC will be here.
+	// 2. A third entry for the target frame PC when the second entry
+	//    is used for skipped inline functions.
+	var pcs [3]uintptr
 	n := runtime.Callers(skip+1, pcs[:])
+	frames := runtime.CallersFrames(pcs[:n])
+	frame, _ := frames.Next()
+	frame, _ = frames.Next()
 
-	var c Call
-	if n < 2 {
-		return c
+	return Call{
+		frame: frame,
 	}
-
-	c.pcs = pcs
-	return c
 }
 
 // String implements fmt.Stinger. It is equivalent to fmt.Sprintf("%v", c).
@@ -48,7 +58,7 @@ func (c Call) String() string {
 // MarshalText implements encoding.TextMarshaler. It formats the Call the same
 // as fmt.Sprintf("%v", c).
 func (c Call) MarshalText() ([]byte, error) {
-	if c.pcs[0] == 0 {
+	if c.frame == (runtime.Frame{}) {
 		return nil, ErrNoFunc
 	}
 
@@ -78,31 +88,19 @@ var ErrNoFunc = errors.New("no call stack information")
 //    %+v   equivalent to %+s:%d
 //    %#v   equivalent to %#s:%d
 func (c Call) Format(s fmt.State, verb rune) {
-	frames := runtime.CallersFrames(c.pcs[:])
-	head, _ := frames.Next()
-	frame, _ := frames.Next()
-
-	if head.Function == "runtime.signpanic" {
-		frame, _ = frames.Next()
-	}
-
-	format(&frame, s, verb)
-}
-
-func format(frame *runtime.Frame, s fmt.State, verb rune) {
-	if frame.Func == nil {
+	if c.frame == (runtime.Frame{}) {
 		fmt.Fprintf(s, "%%!%c(NOFUNC)", verb)
 		return
 	}
 
 	switch verb {
 	case 's', 'v':
-		file, line := frame.File, frame.Line
+		file := c.frame.File
 		switch {
 		case s.Flag('#'):
 			// done
 		case s.Flag('+'):
-			file = file[pkgIndex(file, frame.Function):]
+			file = file[pkgIndex(file, c.frame.Function):]
 		default:
 			const sep = "/"
 			if i := strings.LastIndex(file, sep); i != -1 {
@@ -112,16 +110,15 @@ func format(frame *runtime.Frame, s fmt.State, verb rune) {
 		io.WriteString(s, file)
 		if verb == 'v' {
 			buf := [7]byte{':'}
-			s.Write(strconv.AppendInt(buf[:1], int64(line), 10))
+			s.Write(strconv.AppendInt(buf[:1], int64(c.frame.Line), 10))
 		}
 
 	case 'd':
-		line := frame.Line
 		buf := [6]byte{}
-		s.Write(strconv.AppendInt(buf[:0], int64(line), 10))
+		s.Write(strconv.AppendInt(buf[:0], int64(c.frame.Line), 10))
 
 	case 'k':
-		name := frame.Function
+		name := c.frame.Function
 		const pathSep = "/"
 		start, end := 0, len(name)
 		if i := strings.LastIndex(name, pathSep); i != -1 {
@@ -137,7 +134,7 @@ func format(frame *runtime.Frame, s fmt.State, verb rune) {
 		io.WriteString(s, name[start:end])
 
 	case 'n':
-		name := frame.Function
+		name := c.frame.Function
 		if !s.Flag('+') {
 			const pathSep = "/"
 			if i := strings.LastIndex(name, pathSep); i != -1 {
@@ -152,43 +149,17 @@ func format(frame *runtime.Frame, s fmt.State, verb rune) {
 	}
 }
 
+// Frame returns the call frame infomation for the Call.
+func (c Call) Frame() runtime.Frame {
+	return c.frame
+}
+
 // PC returns the program counter for this call frame; multiple frames may
 // have the same PC value.
+//
+// Deprecated: Use Call.Frame instead.
 func (c Call) PC() uintptr {
-	return c.pcs[1]
-}
-
-func (c Call) frame() *runtime.Frame {
-	frames := runtime.CallersFrames(c.pcs[:])
-	frame, _ := frames.Next()
-	frame, _ = frames.Next()
-	return &frame
-}
-
-// name returns the import path qualified name of the function containing the
-// call.
-func (c Call) name() string {
-	frame := c.frame()
-	if frame.Func == nil {
-		return "???"
-	}
-	return frame.Function
-}
-
-func (c Call) file() string {
-	frame := c.frame()
-	if frame.Func == nil {
-		return "???"
-	}
-	return frame.File
-}
-
-func (c Call) line() int {
-	frame := c.frame()
-	if frame.Func == nil {
-		return 0
-	}
-	return frame.Line
+	return c.frame.PC
 }
 
 // CallStack records a sequence of function invocations from a goroutine
@@ -226,31 +197,12 @@ func (cs CallStack) MarshalText() ([]byte, error) {
 // supplied verb and options.
 func (cs CallStack) Format(s fmt.State, verb rune) {
 	s.Write(openBracketBytes)
-
-	var pcs []uintptr
-	for _, c := range cs {
-		pcs = append(pcs, c.pcs[:]...)
-	}
-
-	frames := runtime.CallersFrames(pcs[:])
-
-	initial := true
-	for frame, more := frames.Next(); more; frame, more = frames.Next() {
-		if !initial {
+	for i, pc := range cs {
+		if i > 0 {
 			s.Write(spaceBytes)
 		}
-
-		next, _ := frames.Next()
-		if frame.Function == "runtime.sigpanic" {
-			frame, _ = runtime.CallersFrames([]uintptr{next.PC - 1}).Next()
-		} else {
-			frame = next
-		}
-
-		format(&frame, s, verb)
-		initial = false
+		pc.Format(s, verb)
 	}
-
 	s.Write(closeBracketBytes)
 }
 
@@ -259,12 +211,17 @@ func (cs CallStack) Format(s fmt.State, verb rune) {
 func Trace() CallStack {
 	var pcs [512]uintptr
 	n := runtime.Callers(1, pcs[:])
-	cs := make([]Call, n)
 
-	for i := range pcs[:n] {
-		cs[i] = Call{
-			pcs: [2]uintptr{pcs[i], pcs[i+1]},
-		}
+	frames := runtime.CallersFrames(pcs[:n])
+	cs := make(CallStack, 0, n)
+
+	// Skip extra frame retrieved just to make sure the runtime.sigpanic
+	// special case is handled.
+	frame, more := frames.Next()
+
+	for more {
+		frame, more = frames.Next()
+		cs = append(cs, Call{frame: frame})
 	}
 
 	return cs
@@ -273,7 +230,7 @@ func Trace() CallStack {
 // TrimBelow returns a slice of the CallStack with all entries below c
 // removed.
 func (cs CallStack) TrimBelow(c Call) CallStack {
-	for len(cs) > 0 && cs[0].pcs[1] != c.pcs[1] {
+	for len(cs) > 0 && cs[0] != c {
 		cs = cs[1:]
 	}
 	return cs
@@ -282,7 +239,7 @@ func (cs CallStack) TrimBelow(c Call) CallStack {
 // TrimAbove returns a slice of the CallStack with all entries above c
 // removed.
 func (cs CallStack) TrimAbove(c Call) CallStack {
-	for len(cs) > 0 && cs[len(cs)-1].pcs[1] != c.pcs[1] {
+	for len(cs) > 0 && cs[len(cs)-1] != c {
 		cs = cs[:len(cs)-1]
 	}
 	return cs
@@ -331,12 +288,13 @@ func pkgIndex(file, funcName string) int {
 var runtimePath string
 
 func init() {
-	var pcs [1]uintptr
+	var pcs [3]uintptr
 	runtime.Callers(0, pcs[:])
-	fn := runtime.FuncForPC(pcs[0])
-	file, _ := fn.FileLine(pcs[0])
+	frames := runtime.CallersFrames(pcs[:])
+	frame, _ := frames.Next()
+	file := frame.File
 
-	idx := pkgIndex(file, fn.Name())
+	idx := pkgIndex(frame.File, frame.Function)
 
 	runtimePath = file[:idx]
 	if runtime.GOOS == "windows" {
@@ -345,7 +303,7 @@ func init() {
 }
 
 func inGoroot(c Call) bool {
-	file := c.file()
+	file := c.frame.File
 	if len(file) == 0 || file[0] == '?' {
 		return true
 	}