--- a/stack_test.go
+++ b/stack_test.go
@@ -224,6 +224,18 @@ func TestTrimRuntime(t *testing.T) {
 	}
 }
 
+func TestPanic(t *testing.T) {
+	defer func() {
+		if recover() != nil {
+			trace := stack.Trace().TrimRuntime()
+			fmt.Println(trace)
+		}
+	}()
+
+	var x *uintptr
+	_ = *x
+}
+
 func BenchmarkCallVFmt(b *testing.B) {
 	c := stack.Caller(0)
 	b.ResetTimer()