--- a/stackinternal_test.go
+++ b/stackinternal_test.go
@@ -3,6 +3,9 @@ package stack
 import (
 	"runtime"
 	"testing"
+	"github.com/dkushner/stack"
+	"strings"
+	"strconv"
 )
 
 func TestCaller(t *testing.T) {
@@ -47,11 +50,53 @@ func TestTrace(t *testing.T) {
 
 	cs := fh.labyrinth()
 
-	lines := []int{43, 33, 48}
+	lines := []int{46, 36, 51}
 
 	for i, line := range lines {
 		if got, want := cs[i].line(), line; got != want {
 			t.Errorf("got line[%d] == %v, want line[%d] == %v", i, got, i, want)
 		}
 	}
+}
+
+// Test stack handling originating from a sigpanic.
+func TestTracePanic(t *testing.T) {
+	t.Parallel()
+
+	defer func() {
+		if recover() != nil {
+			trace := stack.Trace().TrimRuntime()
+
+			if len(trace) != 6 {
+				t.Errorf("got len(trace) == %v, want %v", len(trace), 6)
+			}
+
+			// Check frames in this file, the interceding frames are somewhat
+			// platform-dependent.
+			lines := []int64{68, 101}
+
+			var local []int64
+			for _, call := range trace {
+				parts := strings.Split(call.String(), ":")
+				if parts[0] == "stackinternal_test.go" {
+					line, _ := strconv.ParseInt(parts[1], 10, 32)
+					local = append(local, line)
+				}
+			}
+
+			if len(local) != 2 {
+				t.Errorf("expected %v local frames but got %v", 2, len(local))
+			}
+
+			for i, line := range lines {
+				if got, want := local[i], line; got != want {
+					t.Errorf("got line[%d] == %v, want line[%d] == %v", i, got, i, want)
+				}
+			}
+		}
+	}()
+
+	// Initiate a sigpanic.
+	var x *uintptr
+	_ = *x
 }