@ChrisHines 2017-09-15
@dkushner, I believe this PR represents a complete refactoring to use `runtime.CallersFrames`. Care to take a look?